refactor(test): move contracts test to unit root to reflect cross-layer scope
