+++ src/background.ts
@@ -1,8 +1,8 @@
+import { ContextMenuIds, ContextMenuTitles } from './adapters/chrome/context-menu';
 import { createChromeTabAdapter } from './adapters/chrome/tab';
 import { createChromeTabGroupAdapter } from './adapters/chrome/tab-group';
 import { createChromeWindowAdapter } from './adapters/chrome/window';
 import { mergeWindows } from './app/merge-windows';
-import { ContextMenuIds, ContextMenuTitles } from './context-menu';
 
 const createMergeHandler = (incognito: boolean) => async (): Promise<void> => {
 	const deps = {
